INDXParse: Fix null block bug
In testing security index additions to INDXParse, the M57 "Terry, Dec.
11 #2" (last image of series, and Windows Vista) file system includes a
solid block of 0-value bytes at the end of a security index, when
extracted with an up-to-date Sleuth Kit icat.  This patch adds a
tolerance check for such a null block, in any NTFX index.

The "Terry" images are available at:

<http://digitalcorpora.org/corpora/scenarios/m57-patents-scenario>

Signed-off-by: Alex Nelson <ajnelson@cs.ucsc.edu>
Signed-off-by: Justin Tracey <j2tracey@gmail.com>
Signed-off-by: Erik Steggall <esteggall@soe.ucsc.edu>