Apply isort to all python files with pre-commit
The patch starts continuous integration using pre-commit and GitHub
Actions. As pre-commit requires an initial hook for usage, isort is
filling that role.

The patch starts a GitHub Actions workflow by adapting code from
CASE-Utilities-Python.

The patch tests Python 3.8 and 3.11 as they are the extent of current
supported releases.

References:
* https://github.com/williballenthin/INDXParse/issues/43
* https://github.com/williballenthin/INDXParse/issues/44
* https://www.python.org/downloads/
* https://github.com/casework/CASE-Utilities-Python/pull/37

Disclaimer: Participation by NIST in the creation of the documentation
of mentioned software is not intended to imply a recommendation or
endorsement by the National Institute of Standards and Technology, nor
is it intended to imply that any specific software is necessarily the
best available for the purpose.

Co-authored-by: Alex Nelson <alexander.nelson@nist.gov>
Signed-off-by: Sheldon Douglas <sheldon.douglas@nist.gov>