Transition INDXParse installation to using setup.cfg
This patch implements several revisions to the INDXParse project.

# Disclaimer

Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

# Revisions

## Addition of NIST license text

Because of some significant revisions; copying of files previously
produced by NIST; and new availabity of supplemental license file
references, this patch includes NIST licensing text in some affected
files as well as a standalone file.

## Adoption of `setup.cfg` rather than `pyproject.toml`

INDXParse issue #39 suggests a plan to transition the packaging
framework to `pyproject.toml`.  For a few reasons, this patch initiates
the transition away from `setup.py`, but to `setup.cfg` instead.

The first reason is submitter familiarity with `setup.cfg` features and
non-familiarity with `pyproject.toml` features.  Some complexities are
present with current dependencies of INDXParse, as well as certain
code-review tools, planned for adoption in INDXParse Issue 38, that
require a space to store configuration rules to interact with one
another; I'm familiar with how to address these issues with `setup.cfg`,
but not `pyproject.toml`.  (`isort`'s compatibility configuration is
carried over from how it has been found to work in `case-utils`.)

The second reason is `pyproject.toml` requires resources to use which
are not included in the Python standard library before Python 3.11.
With PEP 0680, `tomllib` is available in Python 3.11.  I suggest the EOL
of Python 3.10 (October of 2026) is a comfortable latest end time to
transition to `pyproject.toml`, but have no objection to a transition
before then.

The `setup.cfg` file implemented in the patch was started by copying
`setup.cfg` from `case-utils` 0.11.0 (see references).  Setup
configurations were mostly reduced to those that ported over from
`setup.py`.

## Switching `setup.py` to `setuptools` from `distutils`

`setup.py` is likewise reduced to the form of a file copied from
`case-utils` 0.11.0.  Usage of `setuptools` rather than `distutils` is
consistent with PyPA guidance (see references, `.../tool-recomendations`).

## Changing scripts' entry points

All scripts intended to be exposed via the command-line `$PATH` now have
a `main()` method.

## Changing modules' listing style

I am unfamiliar with the `py_modules` component within `setup.py`.
I guessed the intent was to guarantee the listed modules would always be
available without a specific import.  `__init__.py` imports those
modules with the spelling it uses so this program runs:

```python
import INDXParse
print(INDXParse.__version__)
print(dir(INDXParse))
print(dir(INDXParse.MFT))
print(dir(INDXParse.BinaryParser))
```

Without the `from . import X` statements, that program fails on
attempting to access `INDXParse.MFT`.  I attempted to do the same with
`__all__`, but couldn't quite follow why modules seemed to be
over-exposed, and then on review of pylance-release issue 2574 (found
from a search engine query for "python __all__"), `__all__` seemed to
not be the way to go.

`mypy` reviews that above 4-`print` program and raises no errors.

## Designating `fuse-mft.py` dependency as extra feature

For macOS users, the `fuse-python` package might have issues with
installation into a virtual environment if FUSE is not installed as a
recent version.  An article from The Register is provided in the
references for further context about macOS FUSE challenges.

Rather than fail the entire installation process if there is any issue
with FUSE, this patch designates an optional feature to only install
`fuse-python` if the user requests it at the time of installing
INDXParse.

The `fuse` feature has been tested by installing this package from
source using the following command with an active virtual environment,
though only in Linux:

    pip install --editable .[fuse]

Should this package be posted to PyPI, the pattern for enabling
`fuse-mft.py` would be to install the package like so:

    pip install INDXParse[fuse]

Otherwise, usage of `fuse-mft.py` would fail on attempting to import
FUSE.

The spelling pattern used within `setup.cfg` to accommodate this
optional package is drawn from the PyPA specifications on entry points
(see references); however, pip issue 9726 documents that this is
non-obviously incorrect, but semi-supported.  I've chosen to include
this not-best-practices syntax as a "Stepping stone" along the way to
implementing a `pyproject.toml`, leaving a reminder of an OS-specific
support issue.

## Designating `MFTView.py` dependency as extra feature

As with `fuse-mft.py`, certain conditions trigger the `wxPython` package
to undergo a long-running build (>10min).  This was observed in Ubuntu
20.04 and 22.04 environments.

Also, "Headless" deployments of INDXParse (those without a GUI) are
unlikely to need the `wxPython` package installed.

For these two reasons, `wxPython` is a dependency in the `wx` feature,
instead of the base installation requirements.

## Reorganization of directory structure

INDXparse Issue #40 sets a plan to migrate code.  I'm currently unsure
how closely that plan as illustrated is technically tied to a
`pyproject.toml` plan.  The `setup.cfg` design in this patch works well
enough with a move of all of the library and script `*.py` files into a
directory named similar to the PyPI package name style; but nothing
should prevent a further reorganization to split the modules and scripts
apart, as long as installation continues to work.

## Revisions to `fuse-mft.py`

The `main()` method was implemented taking command-line arguments.  The
syntax has been adjusted to have `main()` take no arguments, and instead
rely on `sys.argv`.  This change is consistent with the `python-fuse`
source tree's examples (see references, `.../xmp.py`).

# Testing

This patch was tested by building a virtual environment and installing
this source tree into it with this command in macOS:

```bash
pip install --editable .
```

In Linux (Ubuntu 20.04 and 22.04), this was run instead, although just
long enough to confirm the syntax was correct for multiple features:

```bash
pip install --editable .[fuse,wx]
```

The short program printing `dir()` of `INDXParse` modules was then run,
seen to not fail; and `mypy` was run against the program, and seen to
not fail.

Curiously, `mypy --strict` also did not fail against that short program,
but it fails with over 1,700 errors when run against the `INDXParse`
directory (that is, the directory containing `__init__.py`).  These type
errors are known to need addressing in future work.

# References

* https://github.com/casework/CASE-Utilities-Python/blob/0.11.0/setup.cfg
* https://github.com/libfuse/python-fuse/blob/master/example/xmp.py
* https://github.com/microsoft/pylance-release/issues/2574
* https://github.com/pypa/pip/issues/9726
* https://github.com/williballenthin/INDXParse/issues/39
* https://github.com/williballenthin/INDXParse/issues/40
* https://packaging.python.org/en/latest/guides/tool-recommendations/
* https://packaging.python.org/en/latest/specifications/entry-points/
* https://peps.python.org/pep-0680/
* https://www.theregister.com/2019/12/16/fuse_macos_closed_source/

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>