Revise previous type annotations from `bytes` to `array.array` and add necessary type annotations
The patch provides revisions to a previous patch that annotated `buf`
and `self._buf` values to be `bytes` rather than `array.array`. After
additional type review, we believe that an `array.array` annotation
is necessary to account for mutative usages of the buffer variables, as
evidenced by struct.pack_into. The `pack_word` function in
`BinaryParser.py` returns a call to the `struct.pack_into()` function
which requires `self._buf` to be a writeable buffer.

The patch adds type annotations to some `declare_field` fields. The
dynamic population of these fields prevents mypy static type review
from passing, creating a need for type annotation.

The patch provides a return type revision for the `unpack_binary`
function, `bytes` to `array.array`.

The patch adds `array.array` type annotations to each of the lsb2
conversion functions' `binary` parameter.

The patch calls the `tobytes()` function on `self._buf` in
`unpack_wstring` to enable the `.decode()` function.

The patch provides a type annotation to `self._declared_fields` in
`BinaryParser.py` in accordance with the docstring.

The patch provides a type annotation to ` self._nodes` in the `MFTTree`
of `MFT.py` in accordance with its usage.

The patch introduces a clarifying style change for some declared
fields. Following a declared field line is either a single, blank line
or a type annotation for the field directly above. The declared fields
received type annotations only insofar as other effects of this patch
required to enable mypy to pass.  The spacing serves as a visual
reminder of what fields are, and aren't yet, typed.

The patch is written to pass mypy static type review when tested with
the following:
```bash
    mypy list_mft.py MFTINDX.py
```

References:
* https://docs.python.org/3/library/array.html#array.array
* https://docs.python.org/3/library/codecs.html#codecs.decode
* https://docs.python.org/3/library/struct.html#struct.pack_into

Disclaimer: Participation by NIST in the creation of the documentation
of mentioned software is not intended to imply a recommendation or
endorsement by the National Institute of Standards and Technology, nor
is it intended to imply that any specific software is necessarily the
best available for the purpose.

Licensing:
Portions of this patch contributed by NIST are governed by the NIST
Software Licensing Statement:

NIST-developed software is provided by NIST as a public service. You may
use, copy, and distribute copies of the software in any medium, provided
that you keep intact this entire notice. You may improve, modify, and
create derivative works of the software or any portion of the software,
and you may copy and distribute such modifications or works. Modified
works should carry a notice stating that you changed the software and
should note the date and nature of any such change. Please explicitly
acknowledge the National Institute of Standards and Technology as the
source of the software.

NIST-developed software is expressly provided "AS IS." NIST MAKES NO
WARRANTY OF ANY KIND, EXPRESS, IMPLIED, IN FACT, OR ARISING BY OPERATION
OF LAW, INCLUDING, WITHOUT LIMITATION, THE IMPLIED WARRANTY OF
MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
DATA ACCURACY. NIST NEITHER REPRESENTS NOR WARRANTS THAT THE OPERATION
OF THE SOFTWARE WILL BE UNINTERRUPTED OR ERROR-FREE, OR THAT ANY DEFECTS
WILL BE CORRECTED. NIST DOES NOT WARRANT OR MAKE ANY REPRESENTATIONS
REGARDING THE USE OF THE SOFTWARE OR THE RESULTS THEREOF, INCLUDING BUT
NOT LIMITED TO THE CORRECTNESS, ACCURACY, RELIABILITY, OR USEFULNESS OF
THE SOFTWARE.

You are solely responsible for determining the appropriateness of using
and distributing the software and you assume all risks associated with
its use, including but not limited to the risks and costs of program
errors, compliance with applicable laws, damage to or loss of data,
programs or equipment, and the unavailability or interruption of
operation. This software is not intended to be used in any situation
where a failure could cause risk of injury or damage to property. The
software developed by NIST employees is not subject to copyright
protection within the United States.

Co-authored-by: Alex Nelson <alexander.nelson@nist.gov>
Signed-off-by: Sheldon Douglas <sheldon.douglas@nist.gov>