Add process notes and hash records for test data retrieval and preparation
This patch adds a Makefile to record retrieval of DFTT Test 7, selected
for copy-permissive licensing (GPL), small size and for being an image
of only an NTFS file system without encompassing partition system.
This selection was discussed on Issue 41.

The GNU Public License (GPL) license file is included in the zip file.

Running `make` or `make check` in the `third_party` directory prepares
the test data, including hash verification of the disk image.

The Makefile's zip-retrieval instructions are provided not for users to
run, but instead for users to re-run if they so desire.  The recipe is
set to fail if the downloaded resource does not yield hashes as
Git-tracked in this directory.

The hashes recorded were drawn from two sources:

1. The DFTT test page, which provided a MD5 of the disk image contained
   in the zip; and
2. My own hash computations of the zip file, which were not provided on
   the DFTT page; and of the disk image, to use a more up to date
   algorithm.

Hashes for each file $x were computed with OpenSSL 3.1.1, as used in the
Makefile's hash verification steps using `awk`.

References:
* https://dftt.sourceforge.net/test7/index.html
* https://github.com/williballenthin/INDXParse/issues/41

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>