Address typing issues reported and revealed by mypy
These issues were reported when adding a `py.typed` file in accordance
with PEP-561 (to be added in a separate patch), and then running
`mypy indxparse` from the top source directory.  Without the `py.typed`
file, no issues were reported when running `mypy indxparse`.

References:
* https://www.python.org/dev/peps/pep-0561/

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>