Designate INDXParse as a typed project
A downstream project exercising INDXParse as an imported package raised
a type-review error under mypy when using a function from
`indxparse.MFT`:

```
error: Skipping analyzing "indxparse.MFT": module is installed, but missing library stubs or py.typed marker  [import]
note: See https://mypy.readthedocs.io/en/stable/running_mypy.html#missing-imports
```

One solution to this issue is to designate INDXParse as a typed package.
In accordance with PEP-561, the empty (in source-lines-of-code count)
file `py.typed` lets the package be considered a typed package.

This patch adds the empty file and exports it as an additional
installation artifact.

References:
* https://www.python.org/dev/peps/pep-0561/

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>