Expand mypy review to indxparse source directory
This is part of addressing the remark in Issue 38 to use mypy for type
checking of the code base.  `mypy --strict` still raises hundreds of
issues.  But, enabling review of the source tree should at least help
prevent regressions.

References:
* https://github.com/williballenthin/INDXParse/issues/38

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>