Avoid variable name recycling
`mypy` assigns a type to a symbol on its first encounter within a
function body.  When the symbol is reused later, even if for a whole new
purpose, the original name assignment is preserved.  If the type is
incongruous, a type error is raised.

Reviewing `MFTINDX.py` with `mypy --strict` showed several instances of
variable re-use (including in `for` loops), with reassignment to
different types.  This patch renames symbols to avoid ambiguity.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>