Type-annotate mft_get_record_buf and callers
One of the methods flagged as needing type annotations by
`mypy --strict indxparse/MFTINDX.py` is `NTFSFile.mft_get_record_buf`.
This method is written in a way that, for one unfamiliar with calling
context, could return a potentially null result.  Review of the code
paths to this function show that, at present, an `array.array` will be
returned.  But, `MFT.py` would require significant rearchitecting
(likely, subclassing `NTFSFile`) to guarantee that.

Rather than return `Optional[array]`, this patch adds a
`raise ValueError` in accordance with one usage in `MFTINDX.py`.
Handling `None` looked possible, but more complex, than a blunt fail,
due to its usage in the recursive path-constructing function
`mft_record_build_path`.

Some of the other type additions in this patch are side effects from
`mypy` flagging needs, and some are side effects from reviewing the call
paths to `mft_get_record_buf`.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>