Add type annotations for MFTINDX.py
With this patch, `MFTINDX.py` passes `mypy --strict`.  A follow-on patch
will add this status to the testing.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>