Sort some class definitions to satisfy type references
A patch immediately following this one needs to address some type issues
found when a call stack passed through
`NTATTR_STANDARD_INDEX_HEADER.slack_entries()`.  A type-symbol
referenced in that method is (before this patch) defined later in the
file.  `mypy` raises an undefined-symbol error in this case, unless the
syntax for deferring a type symbol with a "Forward reference" is used.

Rather than use the "Forward reference" syntax of a quoted type name,
this patch sorts classes so the code ordering satisfies the topological
type-reference ordering.

The only edit done in this patch was moving classes to their earliest
positions needed to satisfy type references.

References:
* https://peps.python.org/pep-0484/#forward-references
* https://peps.python.org/pep-0484/#the-problem-of-forward-declarations

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>