Update list of .py files currently passing mypy --strict
This is in partial satisfaction of adding mypy type checking, noted on
Issue 38.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* https://github.com/williballenthin/INDXParse/issues/38

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>