Type-annotate and address uncovered issues in list_mft.py
With this patch, `list_mft.py` is added to strict type review.

This patch was edited once to catch indxparse-wide issues not found in
the first revision.
(For quicker round-tripping, the first `mypy` call had been disabled.
That missed new call paths being type-reviewed from the type signatures
added for the sake of `list_mft.py`.)

The class `MFT.Cache` received a `type: ignore` annotation to skip
implementing type variables in a dictionary in a class.  This proved to
be a trickier issue to resolve than first expected.

This patch was edited a second time to remove a typing mechanism not
implemented in Python 3.8.

This is in partial satisfaction of Issue 38.

References:
* https://github.com/williballenthin/INDXParse/issues/38

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>