Type-review most of tree_mft.py
This patch adds type signatures to and around `tree_mft.py` that `mypy --strict`
needs to pass, except for the `Mmap` class, deferred to a future patch.

One function reference is also fixed, now reading as
`get_children_nodes`.

`MFTTreeNode.get_children_nodes` uses the forward reference spelling
(quoted symbol name) in its return type.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* https://peps.python.org/pep-0484/#forward-references

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>