Replace ASCII character sets with stdlib built-in
This patch is in response to `flake8` raising a few dozen warnings like
the following:

```
indxparse/get_file_info.py:30:21: W605 invalid escape sequence '\$'
indxparse/get_file_info.py:30:26: W605 invalid escape sequence '\('
```

This patch increases the number of detected ASCII strings, shown in the
Make-tracked output.

Type annotations are also added in this patch to clarify that the
current implementation has differing expectations of the ASCII character
sets, depending on the containing file.

The symbol 'printable' is imported directly because some of the programs
also use the module name `string` as a loop variable name.  This was
also caught by `flake8`.

A follow-on patch will regenerate Make-managed files.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>