Remove unused function FileMap.unpack
This patch is in response to `flake8` raising the following review note:

```
indxparse/FileMap.py:454:23: F821 undefined name 'buffer'
```

Looking at the containing function, `buffer` did not exist, unlike with
the similarly named function `unpack_from`.  Review of the whole
`indxparse` code base showed no usages of `FileMap.unpack`, so this
patch cuts the function and its shimmed imported symbol.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>