Blind function result
This patch is in response to `flake8` raising the following review note:

```
indxparse/fuse-mft.py:241:17: F841 local variable 'record_buf' is assigned to but never used
```

This patch adds some type review to see whether there was a variable
rename accident, but the review appears to show the variables
`record_buf` and `record` do serve distinct purposes.  A note for a
future reviewer is left instead of removing the call populating the
unused variable `record_buf`.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>