Add `flake8` to `pre-commit` review
This patch is in partial satisfaction of Issue 38.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* https://github.com/williballenthin/INDXParse/issues/38

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>