Add MFT file extracted from current test disk image
This patch provides a test file for INDXParse scripts that take an
extracted MFT as input.

This patch was written to add no new software dependencies to the
testing framework; this is why Python's `gunzip` module was used instead
of the `gunzip` command.

Extraction process notes are inlined in `third_party/Makefile`.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>