Review tree_mft.py with mypy --strict
The `Mmap` class in `tree_mft.py` appears to be a redundant and fallen-
behind implementation of the `Mmap` class in `BinaryParser.py`.  This
patch uses the central implementation instead, which happens to enable
`mpy --strict` to pass.

These changes were tested in Pythons 3.8 through 3.12.  No effects were
observed on Make-managed files.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>