Review self._buf type interactions with mmap.mmap
This patch adds type annotations along the flow from some parameters
named `buf` were changed from being typed `array.array` to `mmap.mmap`.
This review seems to show there are no type conflicts along these
specific changes, according to `mypy` as it runs in CI.

These changes were tested in Pythons 3.8 through 3.12.  No effects were
observed on Make-managed files.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>