Replace array.array type annotations with MutableSequence[int]
On trying to run `MFTINDX.py` against the sample disk image in this
repository, an error was raised, with this (trimmed) call stack:

```
  File "<frozen importlib._bootstrap>", line 228, in _call_with_frames_removed
  File "...[snip].../indxparse/MFTINDX.py", line 284, in <module>
    buf: array.array[Any],
TypeError: 'type' object is not subscriptable
```

(Using Python 3.9.)

This type signature was an attempt to satisfy a review note from
`mypy --strict`, that `array.array` is a generic and needed a further
type specialization, e.g. as visible with
`mypy --strict indxparse/BinaryParser.py` (which is known to also raise
other unrelated issues at the moment):

```
error: Missing type parameters for generic type "array"  [type-arg]
```

This type signature specialization was not runtime-tested before
implementing, and turns out to be incorrect.

A StackOverflow post suggested replacing `array.array` in type
signatures with `collections.abc.MutableSequence`, which is also a
generic but can be specialized.

This patch replaces `array.array` (with and without further type
specialization) in type signatures with `MutableSequence[int]`, because
a single member of a `bytes` object and most other `bytes`-like objects
is an integer.  While this is not entirely true for `array.array`, all
of the usage of `array.array` in this code base is of the `"B"` form,
unsigned byte, so `MutableSequence[int]` is appropriate.  Note that
`array.array` is still used for variables' assignment; it is just not
used in type signatures anymore.

A side-effect of switching away from `array.array` in type signatures is
that `MutableSequence` does not satisfy the `Buffer` protocol expected
from `struct.unpack`.  `.tobytes()` is also not guaranteed to be
available.  `mypy` highlights these issues without `--strict`.  The
`bytes()` built-in method is used to cast all values now only guaranteed
to be `MutableSequence[int]`.

The `Block.unpack_guid` method was modified to remove `ord`, which
happens to not be type-compatible with `MutableSequence[int]` when using
`map`.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* https://stackoverflow.com/a/67775675

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>