Fix usage of types.MethodType
`types.MethodType` in Python 3 takes 2 arguments to instantiate, per
line 447 of the current `types.pyi` in `typeshed`.  I'm not sure when
this switched from 3 arguments as written down to 2; using `git blame`
to step back in the history of that line only goes back to `337abed05a`
in 2015, and doesn't show a specific time where the "name" parameter
this patch removes was expected.

This patch removes the third parameter because it raises a runtime
error.

References:
* https://github.com/python/typeshed/blob/21fcd8960f1dae5ec4563dd99860d0918efe5cff/stdlib/types.pyi#L447

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>