Use type signatures to designate some functions as non-mutative
This patch reviews usage of `MutableSequence[int]`, and restricts
functions that should not modify their inputs (e.g.
`BinaryParser.read_dword`) to use a read-only argument signature instead
of the read-write `MutableSequence[int]`.

An initial suggestion of `typing.ByteString` and/or
`collections.abc.Buffer` led to review of their definitions for whether
they are read-only or read-write types.  The end result is that for
"Read-write" function parameters, `MutableSequence[int]` is still used,
and for "Read-only" parameters, `bytes` is used.

The foundations for this decision are that the `bytes` class is
immutable, `bytearray` is mutable, and the suggested types `ByteString`
and `Buffer` end up supporting `bytearray`.  Prior review of
`typeshed`'s documentation on "readable buffers" and "writable buffers"
indicate there is no "read-only Buffer" generic type currently in
Python.

`collections.abc.Buffer` is the union of `bytes`, `bytearray`, and
`memoryview`.  Because the union includes `bytes` and `bytearray`,
`Buffer` can be mutable.  (This definition was found in
`typing_extensions`; source block cited in references.)

`memoryview` may reference an object that supports the buffer protocol.
This includes `bytes` and `bytearray`.  (This mutability/immutability
detail is documented in both 3.8 and 3.12; 3.8 is cited below.)

`typing.ByteString` is defined as the same union as
`collections.abc.Buffer`, so it can be mutable.

Given all the above, this patch implements "read-only" buffer
restrictions on function parameters with the `bytes` type.

A consequence of using types to restrict read-only vs. read-write is
that several call paths to the read-only functions start from read-write
buffers.  This patch handles this by using `bytes` to cast slices of the
mutable buffers, done to satisfy type restrictions and to avoid casting
larger buffers than necessary.

No effects were observed on Make-managed files.

References:
* https://docs.python.org/3/library/stdtypes.html#bytearray-objects
* https://docs.python.org/3/library/stdtypes.html#bytes-objects
* https://docs.python.org/3.8/library/stdtypes.html#memoryview
* https://github.com/python/typeshed/blob/21fcd8960f1dae5ec4563dd99860d0918efe5cff/stdlib/_typeshed/__init__.pyi#L239-L247
* https://github.com/python/typing_extensions/blob/04f98954ba63a5e8a09c12171be24785298276b6/src/typing_extensions.py#L2522-L2548

Requested-by: Willi Ballenthin <wballenthin@google.com>
Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>